% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{add_suggested_packages}
\alias{add_suggested_packages}
\title{Add suggested pkgs to DESCRIPTION}
\usage{
add_suggested_packages()
}
\value{
Edit the current Suggests DESCRIPTION fields
}
\description{
These pkgs are required to perform the CICD job
}
\keyword{internal}
