% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-app.R
\name{audit_app_docker}
\alias{audit_app_docker}
\title{Run Shiny app validation tools in your project locally with docker}
\usage{
audit_app_docker(
  host_cache = get_renv_cache_path(),
  container_cache = "/root/.renv/cache",
  shinyValidator_tag = NULL,
  port = 80,
  open = TRUE,
  update = FALSE,
  ...
)
}
\arguments{
\item{host_cache}{renv host cache.}

\item{container_cache}{renv cache location inside the container.
Default to "renv/cache".}

\item{shinyValidator_tag}{Allow to get specific flavor of shinyValidator.}

\item{port}{Port where apache2 server will serve the audit HTML report.}

\item{open}{Whether to browse to the apache2 local server url.}

\item{update}{Whether to only update the image. Default to FALSE. Set to TRUE
if you already have a container existing and only want to update the underlying image.}

\item{...}{Parameters to pass to \link{audit_app}.}
}
\description{
Run all specified tools and requirements to validate Shiny apps project.
}
