% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crash-test.R
\name{call_gremlins}
\alias{call_gremlins}
\title{Inject and call gremlins}
\usage{
call_gremlins(headless_app, screenshot = TRUE, path)
}
\arguments{
\item{screenshot}{Whether to take screenshot. Defaults to TRUE. Only works
if headles_actions is NULL.}

\item{path}{Screenshot path.}
}
\description{
Internally required by \link{run_monkey_test}. Tries
to workaround any proxy settings
}
\keyword{internal}
