% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-app.R
\name{check_audit_requirements}
\alias{check_audit_requirements}
\title{Checks to run before running the audit tools}
\usage{
check_audit_requirements()
}
\value{
Error if any of the requirement is not met.
}
\description{
Useful for \link{audit_app}. Briefly, we expect the user
to have shinycannon and Chrome installed since they are required
for loadtest, headless testing ...
}
\keyword{internal}
