% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-dependencies.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Check package dependencies}
\usage{
check_dependencies(r_version = NULL, locked_deps = NULL)
}
\arguments{
\item{r_version}{R version supported by your IT.}

\item{locked_deps}{List of packages supported by your IT. For instance you
can pass a dataframe like
\code{available.packages(repos = "https://cran.microsoft.com/snapshot/2017-01-19/")},
or read a csv file with the same structure.}
}
\value{
An HTML tag
}
\description{
Check package dependencies against a list of
predefined packages. This is useful in locked
environments where package versions are restricted.
}
