% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{check_if_validator_installed}
\alias{check_if_validator_installed}
\title{Checks if the validator is already installed}
\usage{
check_if_validator_installed(cicd_platform)
}
\arguments{
\item{cicd_platform}{CI/CD engine. GitLab or GitHub Actions.}
}
\value{
Boolean.
}
\description{
Checks if the validator is already installed
}
\keyword{internal}
