% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-quality.R
\name{check_package}
\alias{check_package}
\title{Build and check package}
\usage{
check_package(cran = FALSE, vignettes = FALSE, error_on = "never")
}
\arguments{
\item{cran}{Whether to apply as CRAN check. Defaults to FALSE.}

\item{vignettes}{Whether to build vignettes. Defaults to FALSE.}

\item{error_on}{When to raise an error. Possible choices:
\code{c("never", "error", "warning", "note")}. Defaults to never.}
}
\value{
Build and check package. Results are inserted into the
main HTML report.
}
\description{
Build and check package
}
