% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{check_requirements}
\alias{check_requirements}
\title{Checks to run before}
\usage{
check_requirements()
}
\value{
Error if any if the requirement is not met.
}
\description{
Useful for \link{use_validator}
}
\keyword{internal}
