% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{check_setup_requirements}
\alias{check_setup_requirements}
\title{Checks to run before initializing the shinyValidator template}
\usage{
check_setup_requirements(cicd_platform)
}
\arguments{
\item{cicd_platform}{CI/CD engine. GitLab or GitHub Actions.}
}
\value{
Error if any of the requirement is not met.
}
\description{
Useful for \link{use_validator}
}
\keyword{internal}
