% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-callr.R
\name{cleanup_on_exit}
\alias{cleanup_on_exit}
\title{Kill ports on exit}
\usage{
cleanup_on_exit(bg_app, chrome, recorder = NULL)
}
\arguments{
\item{bg_app}{Shiny app background process.}

\item{chrome}{Chrome background process.}

\item{recorder}{Recorder app background process. For shinyloadtest...}
}
\description{
Avoid to have port taken when audit fails
for any reasons
}
\keyword{internal}
