% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-quality.R
\name{create_message_div}
\alias{create_message_div}
\title{Create message container}
\usage{
create_message_div(el)
}
\arguments{
\item{el}{Element such as list of errors, warnings...}
}
\description{
Useful to display R CMD check elements like errors
}
\keyword{internal}
