% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-outputs.R
\name{create_output_tab}
\alias{create_output_tab}
\title{Create output HTML tab}
\usage{
create_output_tab(done)
}
\arguments{
\item{done}{Boolean. Internal to \link{validate_outputs}.}
}
\value{
A shiny HTML tag
}
\description{
Useful for \link{validate_outputs}.
}
\keyword{internal}
