% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-report.R
\name{create_report_steps}
\alias{create_report_steps}
\title{Helper to process audit steps}
\usage{
create_report_steps(
  output_validation = TRUE,
  coverage = TRUE,
  load_testing = TRUE,
  profile_code = TRUE,
  check_reactivity = TRUE,
  flow = TRUE
)
}
\value{
A list composed of dynamic and static steps.
}
\description{
Useful for \link{create_audit_report}.
}
\keyword{internal}
