% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-report.R
\name{create_report_tabs}
\alias{create_report_tabs}
\title{Create a tab menu for HTML report}
\usage{
create_report_tabs(
  output_validation = TRUE,
  coverage = TRUE,
  load_testing = TRUE,
  profile_code = TRUE,
  check_reactivity = TRUE,
  flow = TRUE,
  ...
)
}
\arguments{
\item{...}{To pass extra parameters to \link{edit_html_report}.}
}
\value{
A tab menu used to navigate through \link{create_tab_content}
elements.
}
\description{
Create a tab menu for HTML report
}
\keyword{internal}
