% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-report.R
\name{create_static_tabs}
\alias{create_static_tabs}
\title{Create static tabs}
\usage{
create_static_tabs(static_steps)
}
\arguments{
\item{static_steps}{Provided by \link{create_report_steps}.}
}
\value{
A shiny html tag
}
\description{
Useful in \link{create_audit_report}.
}
\keyword{internal}
