% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-report.R
\name{create_tab_content}
\alias{create_tab_content}
\title{Create a tab content for HTML report}
\usage{
create_tab_content(..., tab_name, title)
}
\arguments{
\item{...}{Tab content.}

\item{tab_name}{Unique tab id. Must match that of \link{create_audit_report}.}

\item{title}{Tab title.}
}
\value{
An HTML tag representing a tab element content.
}
\description{
Used by \link{check_package}, ...
}
\keyword{internal}
