% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-report.R
\name{create_tabs_menu}
\alias{create_tabs_menu}
\title{Create a tab menu}
\usage{
create_tabs_menu(steps, pkg_name, pkg_version)
}
\arguments{
\item{steps}{Audit steps: coverage, output validation, load testing, ...}

\item{pkg_name}{Audited package name.}

\item{pkg_version}{Audited package version.}
}
\value{
A shiny HTML tag
}
\description{
Useful inside \link{create_audit_report}.
}
\keyword{internal}
