% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-report.R
\name{edit_html_report}
\alias{edit_html_report}
\title{Edit HTML validation report}
\usage{
edit_html_report(..., source = path, path = "public/index.html")
}
\arguments{
\item{...}{Allow for parameter expansion.}

\item{source}{Document to edit. Defaults to path to overwrite the
existing report. Set to NULL to fallback to the shinyValidator original
template.}

\item{path}{Where to write. Default to the global report in /public.}
}
\value{
Modified HTML report in public dir.
}
\description{
Useful when items are programmatically generated
and the HTML output cannot be determined in advance.
}
\keyword{internal}
