% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{find_branch}
\alias{find_branch}
\title{Find a git branch}
\usage{
find_branch(name)
}
\arguments{
\item{name}{Branch name.}
}
\value{
Branch name if found
}
\description{
Find a git branch
}
\keyword{internal}
