% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{find_main_branch}
\alias{find_main_branch}
\title{Find the main branch}
\usage{
find_main_branch()
}
\value{
master or main depending on the initial configuration.
}
\description{
Find the main branch
}
\keyword{internal}
