% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-app.R
\name{get_renv_cache_path}
\alias{get_renv_cache_path}
\title{Get renv cache location}
\usage{
get_renv_cache_path()
}
\value{
Character with renv path cache location.
}
\description{
Specific to OS.
Taken from \url{https://rstudio.github.io/renv/articles/renv.html#cache}.
}
\keyword{internal}
