% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{has_git_remote}
\alias{has_git_remote}
\title{Check if current repo has remote part}
\usage{
has_git_remote()
}
\value{
Boolean.
}
\description{
Check if current repo has remote part
}
\keyword{internal}
