% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{initialize_gh_pages}
\alias{initialize_gh_pages}
\title{Checks if gh_pages branch exists}
\usage{
initialize_gh_pages(cicd_platform)
}
\arguments{
\item{cicd_platform}{CI/CD engine. GitLab or GitHub Actions.
When gitlab-docker is selected, we leverage docker executor.
The provided CI/CD template will pull docker image from
\url{https://hub.docker.com/repository/docker/divadnojnarg/shinyvalidator-docker},
which provides installation of R, shinycannon, Chrome, necessary to
run the pipeline without issue.}
}
\description{
Creates gh_pages branch if not
}
\keyword{internal}
