% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-report.R
\name{inject_js_helpers}
\alias{inject_js_helpers}
\title{Inject JS code}
\usage{
inject_js_helpers(steps)
}
\arguments{
\item{steps}{CI/CD steps}
}
\value{
An HTML tag containing the script to be inserted in the report.
}
\description{
Necessary to control the report interactivity
}
\keyword{internal}
