% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-app.R
\name{is_docker_installed}
\alias{is_docker_installed}
\title{Checks if docker is installed on the local machine}
\usage{
is_docker_installed()
}
\value{
Boolean.
}
\description{
Checks if docker is installed on the local machine
}
\keyword{internal}
