% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{is_git_repository}
\alias{is_git_repository}
\title{Check if git is initialized locally}
\usage{
is_git_repository()
}
\value{
Boolean.
}
\description{
Check if git is initialized locally
}
\keyword{internal}
