% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-quality.R
\name{lint_code}
\alias{lint_code}
\title{Lint R code within project}
\usage{
lint_code(paths = "R", tolerance = 0)
}
\arguments{
\item{paths}{Paths to check.}

\item{tolerance}{Errors to allow. Default to 0.}
}
\description{
Uses the \code{{lintr}} package to check all R sources in provided paths.
To fine tune the lintr behavior, edit the \code{.lintr} file.
}
