% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactlog.R
\name{process_reactlog}
\alias{process_reactlog}
\title{Extract and process reactlog}
\usage{
process_reactlog(app)
}
\arguments{
\item{app}{Background app.}
}
\description{
Used by \link{upload_reactlog}.
}
\keyword{internal}
