% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{process_scope}
\alias{process_scope}
\title{Process scope for project}
\usage{
process_scope(scope)
}
\arguments{
\item{scope}{Current project scope}
}
\value{
Reassign parms based on the scope
}
\description{
Process scope for project
}
\keyword{internal}
