% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_app}
\alias{profile_app}
\title{Profile a shiny app}
\usage{
profile_app(timeout = 5, headless_actions = NULL)
}
\arguments{
\item{timeout}{Time to wait after starting the subprocess (s). Useful is you know
how much time the app takes to load.}

\item{headless_actions}{Custom code passed as a string to manipulate the app with headless
web browser, for instance
\code{"headless_app$set_inputs(obs = 200); headless_app$run_js('1+1');"}.
If NULL, the app will undergo a classic random Monkey test session.}
}
\value{
Write a .Rprof file to be reused by CI/CD to publish the report on GitLab pages
}
\description{
Start Shiny + profvis in the background. Chrome connects
to the app and then closes to interrupt profvis.
The profile report is then saved and exported in the public
folder needed for CI/CD.
}
