% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-callr.R
\name{profile_bg}
\alias{profile_bg}
\title{Start Shiny + profvis recorder in the background}
\usage{
profile_bg(...)
}
\arguments{
\item{...}{Pass extra parameters to run_app_audit. This is useful
if you work with packages like golem.}
}
\description{
Required by \link{start_r_bg}.
}
\keyword{internal}
