% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-callr.R
\name{reactlog_bg}
\alias{reactlog_bg}
\title{Start run_app as background process}
\usage{
reactlog_bg(...)
}
\arguments{
\item{...}{Pass extra parameters to run_app. This is useful
if you work with packages like golem.}
}
\description{
Also enables reactlog.
}
\details{
Required by \link{start_r_bg}.
}
\keyword{internal}
