% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-callr.R
\name{recorder_bg}
\alias{recorder_bg}
\title{Start shinyloadtest recorder in the background}
\usage{
recorder_bg(shiny_port)
}
\arguments{
\item{shiny_port}{Port where runs the shiny apps. This is automatically
detected in \link{start_r_bg}.}
}
\description{
Required by \link{start_r_bg}.
}
\keyword{internal}
