% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crash-test.R
\name{run_crash_test}
\alias{run_crash_test}
\title{Crash test for Shiny app}
\usage{
run_crash_test(
  headless_actions = NULL,
  timeout = NULL,
  port = httpuv::randomPort(max = 3500),
  ...
)
}
\arguments{
\item{headless_actions}{Custom code passed as an expression to manipulate the app with headless
web browser, for instance
\code{app$set_inputs(obs = 200); app$run_js('1+1');}.
See \url{https://rstudio.github.io/shinytest2/reference/AppDriver.html} to
get all available methods.
If NULL, the app will undergo a classic random Monkey test session, taking a screenshot
right after loading and after the monkey test. We do not recommend running a monkey test
if your app is pointing to a database, unless the pipeline CI/CD runs in a dedicated test environment.}

\item{timeout}{Time to wait after starting the subprocess (s). Useful is you know
how much time the app takes to load. Defaults to 10 seconds locally and 20 seconds
on CI/CD.}

\item{port}{Port to start the app, default to \code{httpuv::randomPort()}.}

\item{...}{Pass extra parameters to run_app_audit. This is useful
if you work with packages like golem.}
}
\value{
Errors if any of the step fails.
}
\description{
Function required to perform
crash test for a Shiny app. Test whether the app starts and
whether it stays alive after a series of random clicks (monkey test)
or targeted actions.
}
\note{
If the Shiny app takes time to load, you may pass load_timeout
parameter with duration in ms. For instance load_timeout = 5000,
would wait 5 seconds.
}
\examples{
\dontrun{
 # Assuming run_app_audit is defined ...
 run_crash_test(
   headless_actions = {
    app$set_inputs(obs = 1000)
    app$get_screenshot("./plop.png")
   }
 )
}
}
