% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crash-test.R
\name{run_monkey_test}
\alias{run_monkey_test}
\title{Perform basic monkey testing}
\usage{
run_monkey_test(headless_app, headless_actions, screenshot = TRUE)
}
\arguments{
\item{headless_app}{Headless app R6 instance.}

\item{headless_actions}{Custom code passed as a string to manipulate the app with headless
web browser, for instance
\code{"headless_app$set_inputs(obs = 200); headless_app$run_js('1+1');"}.
If NULL, the app will undergo a classic random Monkey test session.}

\item{screenshot}{Whether to take screenshot. Defaults to TRUE.}
}
\description{
Internally required by \link{record_app}, \link{run_crash_test}, ... after
the headless connection is opened.
}
\keyword{internal}
