% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyValidator-package.R
\docType{package}
\name{shinyValidator-package}
\alias{shinyValidator}
\alias{shinyValidator-package}
\title{shinyValidator: Audit Shiny apps in few steps}
\description{
Audit your Shiny apps at each commit. 
    Multiple levels of testings are offered: startup and crash tests, 
    performance tests (load test and global code profiling), reactivity audit as well as output tests.
    All results are gathered in an HTML report uploaded and available to everyone 
    on any CI/CD plaform or RStudio Connect.
}
\author{
\strong{Maintainer}: David Granjon \email{david.granjon@novartis.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Other contributors:
\itemize{
  \item Posit/RStudio (find_chrome related functions.) [copyright holder]
}

}
\keyword{internal}
