% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-callr.R
\name{start_app_with_parms}
\alias{start_app_with_parms}
\title{Start Shiny app with list of parameters}
\usage{
start_app_with_parms(...)
}
\arguments{
\item{...}{Pass extra parameters to run_app. This is useful
if you work with packages like golem.}
}
\value{
Starts a Shiny app.
}
\description{
If not parms are provided, the app is started as usual. Otherwise,
the app is called with \code{do.call} with the provided set of parameters.
Useful for \link{shiny_bg}, \link{profile_bg} and \link{reactlog_bg}.
}
\keyword{internal}
