% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-callr.R
\name{start_r_bg}
\alias{start_r_bg}
\title{Start background R process}
\usage{
start_r_bg(fun, ...)
}
\arguments{
\item{fun}{Passed to \link[callr]{r_bg}.}

\item{...}{Pass extra parameters to run_app. This is useful
if you work with packages like golem.}
}
\value{
Process or error
}
\description{
Start process in the background. Required by
\link{record_app}, ...
}
\keyword{internal}
