% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-outputs.R
\name{validate_outputs}
\alias{validate_outputs}
\title{Validate plot outputs}
\usage{
validate_outputs()
}
\value{
For each snapshot folder found in tests/testthat/_snaps
save a standalone htmlwidget html page to be included in the final report.
}
\description{
Validate plot outputs
}
