% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-callr.R
\name{wait_for_app_action}
\alias{wait_for_app_action}
\title{Wait for application action}
\usage{
wait_for_app_action(action = c("start", "stop"), port)
}
\arguments{
\item{action}{Either "start" or "stop".}

\item{port}{Running port to ping.}
}
\description{
Ping the application port. If action is start,
loop until the port is available, and inversely if action is stop.
}
\keyword{internal}
