% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit-app.R
\name{audit_app}
\alias{audit_app}
\title{Run Shiny app validation tools in your project}
\usage{
audit_app(
  cran = FALSE,
  vignettes = FALSE,
  error_on = "never",
  timeout = NULL,
  headless_actions = NULL,
  workers = 5,
  scope = c("manual", "DMC", "POC"),
  output_validation = FALSE,
  coverage = TRUE,
  load_testing = TRUE,
  profile_code = TRUE,
  check_reactivity = TRUE,
  flow = FALSE,
  ...
)
}
\arguments{
\item{cran}{Whether to apply as CRAN check. Defaults to FALSE.}

\item{vignettes}{Whether to build vignettes. Defaults to FALSE.}

\item{error_on}{When to raise an error. Possible choices:
\code{c("never", "error", "warning", "note")}. Defaults to never.}

\item{timeout}{Time to wait after starting the subprocess (s). Useful is you know
how much time the app takes to load. Defaults to 10 seconds locally and 20 seconds
on CI/CD.}

\item{headless_actions}{Custom code passed as a string to manipulate the app with headless
web browser, for instance
\code{"headless_app$set_inputs(obs = 200); headless_app$run_js('1+1');"}.
If NULL, the app will undergo a classic random Monkey test session.}

\item{workers}{Number of workers for shinycannon. Default to 5.}

\item{scope}{Project scope. Accepted values \code{c("manual", "DMC", "POC")}.}

\item{output_validation}{Whether to compare output snapshots for
plots and htmlwidgets. Default to TRUE.}

\item{coverage}{Whether to perform coverage report. Default to TRUE.}

\item{load_testing}{Whether to perform load test. Default to TRUE.}

\item{profile_code}{Whether to profile R code. Default to TRUE.}

\item{check_reactivity}{Whether to check reactivity log. Default to TRUE.}

\item{flow}{Whether to display project overview. Default to TRUE.}

\item{...}{Pass extra parameters to run_app. This is useful
if you work with packages like golem.}
}
\description{
Run all specified tools and requirements to validate Shiny apps project.
}
