% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{check_if_validator_installed}
\alias{check_if_validator_installed}
\title{Checks if the validator is already installed}
\usage{
check_if_validator_installed(cicd_platform)
}
\arguments{
\item{cicd_platform}{CI/CD engine. GitLab or GitHub Actions.
When gitlab-docker is selected, we leverage docker executor.
The provided CI/CD template will pull docker image from
\url{https://hub.docker.com/repository/docker/divadnojnarg/shinyvalidator-docker},
which provides installation of R, shinycannon, Chrome, necessary to
run the pipeline without issue.}
}
\value{
Boolean.
}
\description{
Checks if the validator is already installed
}
\keyword{internal}
