% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{check_setup_requirements}
\alias{check_setup_requirements}
\title{Checks to run before initializing the shinyValidator template}
\usage{
check_setup_requirements(cicd_platform)
}
\arguments{
\item{cicd_platform}{CI/CD engine. GitLab or GitHub Actions.
When gitlab-docker is selected, we leverage docker executor.
The provided CI/CD template will pull docker image from
\url{https://hub.docker.com/repository/docker/divadnojnarg/shinyvalidator-docker},
which provides installation of R, shinycannon, Chrome, necessary to
run the pipeline without issue. This is typically the easiest setup since
the classic GitLab CI/CD templates assumes that your GitLab runner has
everything installed, which is not necessarily the case.}
}
\value{
Error if any of the requirement is not met.
}
\description{
Useful for \link{use_validator}
}
\keyword{internal}
