% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{edit_buildignore}
\alias{edit_buildignore}
\title{Edit .Rbuildignore file}
\usage{
edit_buildignore()
}
\value{
Edit existing .Rbuidignore file with additional entries
}
\description{
Edit .Rbuildignore file
}
\keyword{internal}
