% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{edit_buildignore}
\alias{edit_buildignore}
\title{Edit .Rbuildignore file}
\usage{
edit_buildignore(cicd_platform, local)
}
\arguments{
\item{cicd_platform}{CI/CD engine. GitLab or GitHub Actions.
When gitlab-docker is selected, we leverage docker executor.
The provided CI/CD template will pull docker image from
\url{https://hub.docker.com/repository/docker/divadnojnarg/shinyvalidator-docker},
which provides installation of R, shinycannon, Chrome, necessary to
run the pipeline without issue. This is typically the easiest setup since
the classic GitLab CI/CD templates assumes that your GitLab runner has
everything installed, which is not necessarily the case.}

\item{local}{Whether to setup the project for local audit with docker.
Default to FALSE.}
}
\value{
Edit existing .Rbuidignore file with additional entries
}
\description{
Edit .Rbuildignore file
}
\keyword{internal}
