% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-setup.R
\name{initialize_cicd}
\alias{initialize_cicd}
\title{Initialize CI/CD template}
\usage{
initialize_cicd()
}
\value{
A yml file with CI/CD steps.
}
\description{
Initialize CI/CD template
}
\keyword{internal}
