% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_app}
\alias{profile_app}
\title{Profile a shiny app}
\usage{
profile_app(headless_actions = NULL, timeout = NULL, ...)
}
\arguments{
\item{headless_actions}{Custom code passed as an expression to manipulate the app with headless
web browser, for instance
\code{app$set_inputs(obs = 200); app$run_js('1+1');}.
See \link{https://rstudio.github.io/shinytest2/reference/AppDriver.html} to
get all available methods.
If NULL, the app will undergo a classic random Monkey test session, taking a screenshot
right after loading and after the monkey test.}

\item{timeout}{Time to wait after starting the subprocess (s). Useful is you know
how much time the app takes to load. Defaults to 10 seconds locally and 20 seconds
on CI/CD.}

\item{...}{Pass extra parameters to run_app. This is useful
if you work with packages like golem.}
}
\value{
Write a .Rprof file to be reused by CI/CD to publish the report on GitLab pages
}
\description{
Start Shiny + profvis in the background. Chrome connects
to the app and then closes to interrupt profvis.
The profile report is then saved and exported in the public
folder needed for CI/CD.
}
