% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-test.R
\name{record_app}
\alias{record_app}
\title{Record a Shiny session}
\usage{
record_app(timeout = 5, headless_actions = NULL, workers = 5, ...)
}
\arguments{
\item{timeout}{Time to wait after starting the subprocess (s). Useful is you know
how much time the app takes to load.}

\item{headless_actions}{Custom code passed as a string to manipulate the app with headless
web browser, for instance
\code{"headless_app$set_inputs(obs = 200); headless_app$run_js('1+1');"}.
If NULL, the app will undergo a classic random Monkey test session.}

\item{workers}{Number of workers for shinycannon. Default to 5.}

\item{...}{Pass extra parameters to run_app. This is useful
if you work with packages like golem.}
}
\description{
Start Shiny app as local R subprocess. Connect the
shinyloadtest recorder and connect Chrome to the recorder.
Manipulate the headless Chrome and close connection. shinycannon
replays
}
