% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crash-test.R
\name{run_monkey_test}
\alias{run_monkey_test}
\title{Perform basic monkey testing}
\usage{
run_monkey_test(app, headless_actions, screenshot = TRUE, path)
}
\arguments{
\item{app}{Headless app R6 instance.}

\item{headless_actions}{Custom code passed as an expression to manipulate the app with headless
web browser, for instance
\code{app$set_inputs(obs = 200); app$run_js('1+1');}.
See \url{https://rstudio.github.io/shinytest2/reference/AppDriver.html} to
get all available methods.
If NULL, the app will undergo a classic random Monkey test session, taking a screenshot
right after loading and after the monkey test. We do not recommend running a monkey test
if your app is pointing to a database, unless the pipeline CI/CD runs in a dedicated test environment.}

\item{screenshot}{Whether to take screenshot. Defaults to TRUE. Only works
if headles_actions is NULL.}

\item{path}{Screenshot path.}
}
\description{
Internally required by \link{record_app}, \link{run_crash_test}, ... after
the headless connection is opened.
}
\keyword{internal}
