% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_validator}
\alias{use_validator}
\title{Include Shiny app validation tools in your project}
\usage{
use_validator(
  cicd_platform = c("github", "gitlab-docker", "gitlab"),
  local = FALSE
)
}
\arguments{
\item{cicd_platform}{CI/CD engine. GitLab or GitHub Actions.
When gitlab-docker is selected, we leverage docker executor.
The provided CI/CD template will pull docker image from
\url{https://hub.docker.com/repository/docker/divadnojnarg/shinyvalidator-docker},
which provides installation of R, shinycannon, Chrome, necessary to
run the pipeline without issue. This is typically the easiest setup since
the classic GitLab CI/CD templates assumes that your GitLab runner has
everything installed, which is not necessarily the case.}

\item{local}{Whether to setup the project for local audit with docker.
Default to FALSE.}
}
\description{
Setup all necessary tools and requirements to validate Shiny apps project.
This function is the first to be called.
}
\details{
By default, the package is checked, built and test are run.
Also, we quickly check if the Shiny application is able to start and run
without crashing.
}
