% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_validator}
\alias{use_validator}
\title{Include Shiny app validation tools in your project}
\usage{
use_validator(cicd_platform = c("gitlab", "github"))
}
\arguments{
\item{cicd_platform}{CI/CD engine. GitLab or GitHub Actions.}
}
\description{
Setup all necessary tools and requirements to validate Shiny apps project.
This function is the first to be called.
}
\details{
By default, the package is checked, built and test are run.
Also, we quickly check if the Shiny application is able to start and run
without crashing.
}
